+++ app/Http/Controllers/PedidoController.php
@@ -6,38 +6,34 @@ use App\Models\Pedido;
 use Illuminate\Http\Request;
 use App\Http\Requests\PedidoRequest;
 use Uspdev\Replicado\Graduacao;
+use App\Service\PedidoStatus;
+use App\Service\Utils;
+use Illuminate\Support\Facades\Gate;
+use Illuminate\Support\Facades\Storage;
 
 class PedidoController extends Controller
 {
     public function index(Request $request)
     {
         $this->authorize('admin');
-        #Campo de busca
-        
-
-        if ($request->buscastatus != null && $request->search != null){
-
-            $pedidos = Pedido::where($this->user->codpes == 'user_id','LIKE',"%{$request->search}%")
-              ->where(Pedido::getStatus(), $request->buscastatus)->paginate(10);
-            
-        } else if(isset($request->search)) {
-            $pedidos = Pedido::where($this->user->codpes == 'user_id','LIKE',"%{$request->search}%")->paginate(10);
+        $pedidos = new Pedido;
+        if(isset($request->buscastatus)) {
+            $pedidos =  $pedidos->where('status',$request->buscastatus);
         }
-            
-        else if(isset($request->buscastatus)){
-            $pedidos = Pedido::where(Pedido::getStatus(), $request->buscastatus)->paginate(10);
-        } else {
-            $pedidos = Pedido::paginate(10);
+
+        if(isset($request->search)) {
+            $pedidos =  $pedidos->where('codpes','LIKE',"%{$request->search}%")
+                                ->orWhere('nome','LIKE',"%{$request->search}%");
         }
 
         return view ('pedidos.index',[
-            'pedidos' => $pedidos
+            'pedidos' => $pedidos->get()
         ]);
     }
 
     public function create()
     {
-        $this->authorize('logado');
+        $this->authorize('grad');
         return view('pedidos.create',[
         'pedido' => new Pedido,
         ]);
@@ -45,34 +41,30 @@ class PedidoController extends Controller
 
     public function store(PedidoRequest $request)
     {
-        $this->authorize('logado');
+        $this->authorize('grad');
         $validated = $request->validated();
         $validated['user_id'] = auth()->user()->id;
         $validated['original_name'] = $request->file('file')->getClientOriginalName();
         $validated['path'] = $request->file('file')->store('.');
         $pedido = Pedido::create($validated);
+        $pedido->save();
+
         request()->session()->flash('alert-info','Cadastro com sucesso');
         return redirect("/pedidos/$pedido->id");
     }
 
-    public function show(Pedido $pedido)
+    public function show(Pedido $pedido, PedidoStatus $stepper)
     {
         $this->authorize('owner',$pedido);
-        $codpes = auth()->user()->codpes;
 
-        $curso = Graduacao::curso($codpes,8);
-        if(!empty($curso)) {
-            $disciplinas = Graduacao::listarDisciplinasGradeCurricular($curso['codcur'], $curso['codhab']);
-        } else {
-            $disciplinas = [];
-        }
+        $stepper->setCurrentStepName($pedido->status);
+        $codpes = auth()->user()->codpes;
 
         return view('pedidos.show',[
             'pedido' => $pedido,
-            'disciplinas' => $disciplinas
+            'disciplinas' => Utils::disciplinas(auth()->user()->codpes),
+            'stepper' => $stepper->render()
         ]);
-
-        
     }
 
     public function edit(Pedido $pedido)
@@ -95,7 +87,34 @@ class PedidoController extends Controller
     public function destroy(Pedido $pedido)
     {
         $this->authorize('owner',$pedido);
+
+        # o aluno só pode deletar enquanto estiver em elaboração
+        if($pedido->status != 'Em elaboração' & !Gate::allows('admin')){
+            request()->session()->flash('alert-danger','Esse pedido não pode ser deletado 
+                porque não está mais em elaboração');
+            return redirect('/meus_pedidos');
+        }
+        
+        Storage::delete($pedido->path);
+        foreach($pedido->disciplinas as $disciplina){
+            Storage::delete($disciplina->path);
+            $disciplina->delete();
+        }
         $pedido->delete();
-        return redirect('/pedidos');
+        return redirect('/');
+    }
+
+    public function meus_pedidos(){
+        $this->authorize('grad');
+        $pedidos = Pedido::where('user_id',auth()->user()->id)->paginate(10);
+        return view('pedidos.meus_pedidos',[
+            'pedidos' => $pedidos
+        ]);
+    }
+
+    public function showfile(Pedido $pedido)
+    {
+        $this->authorize('owner',$pedido);
+        return Storage::download($pedido->path, $pedido->original_name);
     }
 }
